requirements/base.txt
Django==2.1.7
django-dotenv==1.4.2
django-filter==2.1.0
djangorestframework==3.9.1
drf-yasg==1.11.0
flex==6.14.0
pika==0.13.0
Pillow==5.4.1
psycopg2-binary==2.7.7
redis==3.2.0
swagger-spec-validator==2.4.3
django-model-utils==3.1.2
django-searchable-select==1.5.0
django-taggit==1.1.0
